dmaengine: Allow per-engine channel filtering callback
Currently the filtering function is received from platform code inside
platform data, but usually this function is just a generic code
implemented and exported from drivers of particular DMA engines.

Based on this, we can simply move registration of the filtering callback
to DMA engine driver. This will allow for further simplification of DMA
channel handling introduced in further patches.

Signed-off-by: Tomasz Figa <tomasz.figa@gmail.com>